k8s: Support v1 & v1beta1 CRDs
This commit implements support for both v1 and v1beta1 CRDs in Cilium.
v1 CRDs will be used when the K8s version if 1.16 or higher, otherwise
v1beta1 CRDs will be used.

By virtue of the apiserver's ability to interoperate between the two
versions, we can rely on it handle conversions between v1 and v1beta1
without writing any explicit logic in Cilium. We simply only need to
provide v1 and v1beta1 variants when interacting with K8s 1.16 or higher
and 1.15 or lower, respectively.

This commit also ensures we are ready for the removal of v1beta1 CRDs,
which will happen in Kubernetes 1.22 [1].

[1]:
https://github.com/kubernetes/kubernetes/blob/c182a599ee3f02b4b66c6d20caaa05f9a406a10f/staging/src/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1/types.go#L395

Signed-off-by: Chris Tarazi <chris@isovalent.com>